Local double editor. Renamed all references to LocalFloat to LocalDouble

+++ Ceriyo.Core/Scripting/Server/Contracts/ILocalDataMethods.cs
@@ -5,9 +5,9 @@ namespace Ceriyo.Core.Scripting.Server.Contracts
     public interface ILocalDataMethods
     {
         void SetLocalValue(Entity entity, string key, string value);
-        void SetLocalValue(Entity entity, string key, float value);
+        void SetLocalValue(Entity entity, string key, double value);
         string GetLocalString(Entity entity, string key);
-        float GetLocalNumber(Entity entity, string key);
+        double GetLocalNumber(Entity entity, string key);
         void DeleteLocalString(Entity entity, string key);
         void DeleteLocalNumber(Entity entity, string key);
     }
